Renaming ServerCallback to ServerObserver
